The map validator is at 90%, need make flood fill to check if there is a valid path

+++ GNL/get_next_line.h
@@ -25,7 +25,7 @@ char	*ft_strchr(const char *str, int c);
 char	*strjoin_and_free(char *s1, const char *s2);
 char	*extract_line(char *buffer);
 void	*ft_memmove(void *dest, const void *src, size_t n);
-size_t	ft_strlen(const char *str);
+int	ft_strlen(const char *str);
 char	*ft_strdup(const char *str);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
 char	*final_line(char **buffer);
